Update dataset
The GitHub search is now returning incomplete results because it started
exceeding the time limits.¹  As a partial, manual workaround for now, I made this
data refresh more complete by cumulatively combining multiple runs along
with the previous data.  This means any already-known results which have
since disappeared are still included, but I suspect that's few.

    cp community.ndjson tmp/community.ndjson.1

    for i in $(seq 2 10); do
        ./find > tmp/community.ndjson.$i
    done

    sort -u tmp/community.ndjson.* | \
        jq -sc 'sort_by([.repo.full_name, .file]) | .[]' \
        > community.ndjson

This workaround could be automated in the near future and use a
threshold for when to stop instead of a fixed number of iterations (e.g.
stop when less then N new results are found in the latest iteration).

¹ https://docs.github.com/en/rest/reference/search#timeouts-and-incomplete-results