Automate a partial workaround to incomplete search results
As the previous commit noted, GitHub search is now returning incomplete
results because it started exceeding the time limits.¹  Partially work
around this by cumulatively combining multiple search runs into the
existing dataset.  This means that we won't remove results from the
dataset when files are deleted, but I suspect that's a relatively
uncommon occurrence (and we might want to keep historical community
builds in this dataset anyway).

¹ https://docs.github.com/en/rest/reference/search#timeouts-and-incomplete-results